+++ src/server/game/Entities/Player/Player.cpp
@@ -20321,7 +20321,7 @@ void Player::TalkedToCreature(uint32 entry, uint64 guid)
     QuestObjectiveSatisfy(entry, 1, QUEST_OBJECTIVE_TYPE_NPC_INTERACT, guid);
 }
 
-void Player::MoneyChanged(uint32 count)
+void Player::MoneyChanged(uint64 count)
 {
     for (uint8 i = 0; i < MAX_QUEST_LOG_SIZE; ++i)
     {
