+++ src/server/game/Entities/Unit/Unit.h
@@ -1644,7 +1644,7 @@ class Unit : public WorldObject
         bool HealthAbovePctHealed(int32 pct, uint32 heal) const { return uint64(GetHealth()) + uint64(heal) > CountPctFromMaxHealth(pct); }
         float GetHealthPct() const { return GetMaxHealth() ? 100.0f * GetHealth() / GetMaxHealth() : 0.0f; }
         uint32 CountPctFromMaxHealth(int32 pct) const { return CalculatePct(GetMaxHealth(), pct); }
-        uint32 CountPctFromMaxHealth(float p_Percent) const { return CalculatePct((float)GetMaxHealth(), p_Percent); }
+        uint32 CountPctFromMaxHealth(float p_Percent) const { return uint32(CalculatePct((float)GetMaxHealth(), p_Percent)); }
         uint32 CountPctFromCurHealth(int32 pct) const { return CalculatePct(GetHealth(), pct); }
         uint32 CountPctFromMaxMana(int32 pct) const { return CalculatePct(GetMaxPower(POWER_MANA), pct); }
         uint32 CountPctFromCurMana(int32 pct) const { return CalculatePct(GetPower(POWER_MANA), pct); }
@@ -2983,8 +2983,8 @@ namespace JadeCore
             AreaTriggerDurationPctOrderPred(bool ascending = true) : m_ascending(ascending) {}
             bool operator() (const AreaTrigger* a, const AreaTrigger* b) const
             {
-                int32 rA = a->GetDuration() ? float(a->GetDuration()) : 0;
-                int32 rB = b->GetDuration() ? float(b->GetDuration()) : 0;
+                int32 rA = a->GetDuration() ? a->GetDuration() : 0;
+                int32 rB = b->GetDuration() ? b->GetDuration() : 0;
                 return m_ascending ? rA < rB : rA > rB;
             }
         private:
